requirements.txt
Flask
gunicorn
flask_sqlalchemy
sqlalchemy
python-docx
